Make sendAndWait timeout parameter nullable
Updated the sendAndWait method in CopilotSession, Session, and FakeSession to accept a nullable timeout parameter. The default timeout is now taken from the configuration if not provided. Also updated config and documentation to reflect these changes.

+++ config/copilot.php
@@ -80,7 +80,7 @@ return [
     | The default timeout in seconds for sendAndWait operations.
     |
     */
-    'timeout' => env('COPILOT_TIMEOUT', 60),
+    'timeout' => (float) env('COPILOT_TIMEOUT', 60.0),
 
     /*
     |--------------------------------------------------------------------------
